requirements.txt
opencv-python==4.8.1.78
PyQt5==5.15.9
numpy==1.24.3
Pillow==10.0.0
pyyaml==6.0.1
Flask==2.3.3
Werkzeug==2.3.7
